feat: create Components for better readability
